+++ Core/Constants.lua
@@ -44,6 +44,9 @@ addon.OPT_MORE_BUTTONS_VALUES = {
 	-- CORRUPTED = _G.CORRUPTION_TOOLTIP_TITLE,
 	CONTEXT_TOKEN = "Beads and Spherules",
 	PETS = _G.PETS,
+	MOUNTS = _G.MOUNTS,
+	BAGSLOT = _G.BAGSLOT,
+	RECIPE = _G.RECIPE
 }
 
 --[[
@@ -117,6 +120,27 @@ addon.RESPONSE_CODE_GENERATORS = {
 			if addon.BTN_SLOTS[itemEquipLoc] == "AZERITE" then return "AZERITE" end
 		end
 	end,
+
+	   -- Mounts
+   function (_, db, _, _, classID, subClassID)
+      if db.enabledButtons.MOUNTS and classID == Enum.ItemClass.Miscellaneous and subClassID == Enum.ItemMiscellaneousSubclass.Mount then
+         return "MOUNTS"
+      end
+   end,
+
+   -- Bags
+   function(_, db, _, _, classID)
+      if db.enabledButtons.BAGSLOT and classID == Enum.ItemClass.Container then
+         return "BAGSLOT"
+      end
+   end,
+
+   -- Recipies
+	function(_, db, _, _, classID)
+		if db.enabledButtons.RECIPE and classID == Enum.ItemClass.Recipe then
+			return "RECIPE"
+		end
+   end
 }
 
 --- @alias VersionCodes