+++ changelog.md
@@ -1,3 +1,22 @@
+## Changes
+
+### Voting Frame More Info
+
+Now shows the equip location of recently awarded items.
+
+### Export items in session
+
+Added new chat command `/rc export` which will export a csv formatted list of the items currently in session.
+
+### Button Groups
+
+Added button groups for mounts, bags and recipies.
+
+## Bugfixes
+
+- *Items can now again be automatically added to a trade with a player from another realm.*
+- *Fixed issue with items some times being added twice.*
+
 # 3.6.7
 
 ## Bugfixes